move survival module
