+++ foundry/glm/family/__init__.py
@@ -1,2 +1,1 @@
 from .family import Family
-from .survival import CeilingWeibull, SurvivalFamily
